+++ packages/documentation/scripts/build.ts
@@ -26,7 +26,7 @@ const buildIndex = (): void => {
     write: false,
     external: ['typescript']
   })
-  const { path, text } = code.outputFiles[0]!
+  const { path, text } = code.outputFiles[0]
 
   const transformText = (text: string): string => {
     return text