+++ packages/react-example/src/main.ts
@@ -1,9 +1,13 @@
 
-// eslint-disable-next-line import/no-unassigned-import -- import css
+// eslint-disable-next-line import-x/no-unassigned-import -- import css
 import './statics/styles/reset.css'
-// eslint-disable-next-line import/no-unassigned-import -- import css
+
+
+// eslint-disable-next-line import-x/no-unassigned-import -- import css
+
 import 'virtual:uno.css'
-// eslint-disable-next-line import/no-unassigned-import -- import css
+
+// eslint-disable-next-line import-x/no-unassigned-import -- import css
 import './statics/styles/styles.css'
 
 import React from 'react'