+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @we-mobius/eslint-config
 
+## 0.0.4
+
+### Patch Changes
+
+- fix eslint-config error
+
+## 0.0.3
+
+### Patch Changes
+
+- update deps to latest version
+
 ## 0.0.2
 
 ### Patch Changes