+++ packages/react-example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @we-mobius/react-example
 
+## 0.0.3
+
+### Patch Changes
+
+- fix eslint-config error
+
 ## 0.0.2
 
 ### Patch Changes