+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @we-mobius/tsconfig
 
+## 0.0.10
+
+### Patch Changes
+
+- fix eslint-config error
+
+## 0.0.9
+
+### Patch Changes
+
+- update deps to latest version
+
+## 0.0.8
+
+### Patch Changes
+
+- update tsconfigs up to TypeScript 5.4 & update deps to latest
+
 ## 0.0.7
 
 ### Patch Changes